Started game config
